refactor(ui): GR_Canvas is now the G6 canvas; delete the SVG renderer
Consolidates onto a single canvas: the G6 implementation moves into the
canonical graph-canvas.jsx as window.GR_Canvas (function GR_Canvas), the
hand-rolled SVG renderer + its primitives (GR_NodeBox/GR_EdgePath/etc.) are
deleted, and the spike-named graph-g6-canvas.jsx is gone. Consumers
(session-detail run-view, graphs editor) point at window.GR_Canvas; index.html
registers the one file. GR_NODE_SIZE is retained. Live-verified both surfaces
render via the single canvas with no regressions.
