+++ ui/index.html
@@ -73,7 +73,6 @@
 <script type="text/babel" src="components/shared/mobile-tabs.jsx"></script>
 <script type="text/babel" src="components/shared/floating-action.jsx"></script>
 <script type="text/babel" src="components/graph-canvas.jsx"></script>
-<script type="text/babel" src="components/graph-g6-canvas.jsx"></script>
 <script type="text/babel" src="components/mock-data.jsx"></script>
 <script type="text/babel" src="components/predicate-builder.jsx"></script>
